some pre-commit formatting
